The parser does a pre pass to set the name of objects from their scripts.
The #bind command has been added back and it is
now functional, but it simply copies the scripts
from the object it is binding to.

The #text command was added back.

Old execute object commands function was removed.

The :click label was added. This is a special
label like :touch but triggers when an object
is clicked on with the mouse.